Decomposing blacklight generator
Created a blacklight:models generator that is now used by the
blacklight generator.
